Move the convert to blocks option from menu to own toolbar button
