+++ packages/block-editor/src/components/block-settings-menu/block-settings-dropdown.js
@@ -25,7 +25,6 @@ import { serialize } from '@wordpress/blocks';
 import BlockActions from '../block-actions';
 import BlockModeToggle from './block-mode-toggle';
 import BlockHTMLConvertButton from './block-html-convert-button';
-import BlockUnknownConvertButton from './block-unknown-convert-button';
 import __experimentalBlockSettingsMenuFirstItem from './block-settings-menu-first-item';
 import BlockSettingsMenuControls from '../block-settings-menu-controls';
 
@@ -106,11 +105,6 @@ export function BlockSettingsDropdown( {
 								<__experimentalBlockSettingsMenuFirstItem.Slot
 									fillProps={ { onClose } }
 								/>
-								{ count === 1 && (
-									<BlockUnknownConvertButton
-										clientId={ firstBlockClientId }
-									/>
-								) }
 								{ count === 1 && (
 									<BlockHTMLConvertButton
 										clientId={ firstBlockClientId }
